Switch back to using _dt instead of stage_dt.
The latter (as we've computed it) doesn't make sense for explicit methods (it is always zero).

Add more print statements.

The family of ExplicitRK2 methods also don't work with Yaqi's test
problem, but for a different reason: the integral value is accumulated
twice as much (for Heun's method) than it should be because of the
advanceState() call that happens between the stages...