More debugging print statements.
I think this is actually working correctly - you have to remember that
if the subapp is a 2 element mesh, the master app nodes at e.g. x=0.25
could equally get either the x=0 or x=0.5 subapp value transferred
into them... In my output file it looks like the value u=0.5 is being
transferred to the in-between values in all cases.